requirements.txt
asgiref==3.5.1
DatasetCreator==0.1
Django==4.0.4
djangorestframework==3.13.1
googleads==3.8.0
httplib2==0.20.4
joblib==1.1.0
numpy==1.22.3
oauth2client==4.1.3
pandas==1.4.2
praat-parselmouth==0.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2015.7
PyYAML==6.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
stopit==1.1.1
suds-jurko==0.6
threadpoolctl==3.1.0
